requirements.txt
langchain_community 
tiktoken==0.7.0
langchain-openai==0.1.19
langchainhub==0.1.20
chromadb==0.5.3 
langchain==0.2.11
langchain-google-community==1.0.7
langchain-core<0.3.0
PyPDF2==3.0.1
altair==4.0
streamlit==1.12.0
oauth2client<4.0.0
bs4==0.0.2
google-auth-oauthlib==1.2.1